refactor OAuth integration types and move to new types module; remove creds helper and just rely on usage of content item associations find_by
